fix: moved ready_aim_fire method to game class to access computer class
